refactor: 서버 통신 로직을 useRestaurants 커스텀 훅으로 분리
